ml-service/requirements.txt
# ML & Data Science
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# API
flask==3.0.0
flask-cors==4.0.0

# Database
pymysql==1.1.0
cryptography==41.0.7

# Utils
python-dotenv==1.0.0

